[refactor] 객체가 의도한 책임을 하도록 controller 로직 변경
