+++ src/main/java/domain/answer/AnswerType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.answer;
 
 import java.util.Arrays;
